Fixed #2
Images (or other media) are saved in files whose name is built from
the timestamp of the message in Telegram.

This caused issues when several media share the same timestamp (because
they were uploaded together in Telegram).

Now each downloaded file contains a suffix _0, _1,... to differentiate
among the ones with the same timestamp.